Fixed uniform location bug
So apparently uniform indexes and locations are two diffrent things and i confused the two. Thank you to this blog post https://www.alecjacobson.com/weblog/2203.html for helping me fix this issue. Somehow on Linux the indexes and locations would be equal but when i tested it on windows that was no longer the case.